Ignore non-empty directories when cleanup
[PR link](https://github.com/aws/aws-codedeploy-agent/pull/201)
Description of changes

When installing ruby in ubuntu 18.04, ruby version 2.5.x is installed by default.

Before ruby 2.5(i.e., 2.4.3), FileUtils.rmdir silently fails when the target directory is not empty, so the non-empty directories in cleanup file are not removed on deployment, and the deployment succeeds.

But the implementation of FileUtils.rmdir is changed in ruby 2.5 so that it doesn’t rescue Errno::ENOTEMPTY any more, which generates failures on deployment.

This PR adds a rescue from Errno::ENOTEMPTY, thus non-empty directories in cleanup file are not removed for ruby 2.5 and forward, and the deployment does not fail.

Note: There are two other exceptions(Errno::EEXIST, Errno::ENOENT) that can occur in FileUtils.rmdir and have been rescued before ruby 2.5. We don't rescue them here because:

    Errno::EEXIST - File exists: it doesn't occur in our use case. (Frankly I have no idea why this rescue is needed from the first time)
    Errno::ENOENT - No such file or directory: it doesn't occur since we first check File.exist?(@file_path) before FileUtils.rmdir(@file_path) is executed.

cr https://code.amazon.com/reviews/CR-20444660