[Agent 1.1.2] Update version of the agent to 1.1.2
Why is this change needed?
---------
Prior to this change, Agent version was set to 1.1.1, so for the new release we need to set to 1.1.2

How does it address the issue?
---------
This change will update the patch version of the agent, for new release 1.1.2

How was this tested ?
---------
 * Unit Tests : bb release
 * Integration Tests : bb test-integration

Provide links to any relevant tickets(SIM Issue/TT), articles or other resources
---------

cr https://code.amazon.com/reviews/CR-29766806