[Agent 1.1.2] Add unit test to verify unpack_bundle ruby unzip mechanism
Why is this change needed?
---------
Prior to this change, we didn't have a way to test the "overwrite" logic
in ruby unzip during the fallback mechanism of unpack_bundle method in
DownloadBundle command.

How does it address the issue?
---------
This change adds a unit test to verify if the fallback mechanism works
with a debris file in the deployment archive directory.

How was this tested ?
---------
 * Unit Tests : bb
 * Integration Tests : bb test-integration

Provide links to any relevant tickets(SIM Issue/TT), articles or other resources
---------
https://i.amazon.com/P37512648

cr https://code.amazon.com/reviews/CR-30141191