[IMDSv2 support] Update agent code to use IMDSv2
Prior to this change, when retrieving instance metadata, the agent used IMDSv1. There is an escalation to remove all our dependencies on IMDSv1 and update them to use IMDSv2. This is a large ask because of the security vulnerabilities associated with IMDSv1.

This change updates our agent code to use IMDSv2, which uses a session authentication method for retrieving instance metadata.

 * Unit Tests : [Y]
 * Integration Tests : [N]
 * Successfully tested install script and agent by spinning up an ec2 instance, building the rpm file for the agent using RBHAInstaller, and deploying a sample application via CodeDeploy console to that instance.

SIM for this feature request: https://sim.amazon.com/issues/CODEDEPLOY-5447

cr https://code.amazon.com/reviews/CR-30143152