[Agent upgrade 1.2.0] Upgrade agent version to 1.2.0
Prior to this change, the agent version was set to 1.1.3 for the IMDSv2 updates. However, that was a mistake because following the new versioning format, IMDSv2 counts as a new feature, and thus needs to be minor version bump up.

This change updates the agent version to 1.2.0.

 * Unit Tests : [Y]
 * Integration Tests : Manually successfully ran a deployment with this agent version.

cr https://code.amazon.com/reviews/CR-30633503