[chore] Reverting part of the rake update
ProcessManager, a long abandoned project and "dependency" of this
package (the source code is copy/pasted strait in) declare needing rake
though actually makes no use of it. It actually declares rake and
several other packages several times which smells of a bad merge but
that is out of scope.

cr https://code.amazon.com/reviews/CR-37337765